Successful make distribution

+++ minix/drivers/storage/memory/memory.c
@@ -414,7 +414,7 @@ static struct device *m_block_part(devminor_t minor)
 /*===========================================================================*
  *				m_block_transfer			     *
  *===========================================================================*/
-static int m_block_transfer(
+static ssize_t m_block_transfer(
   devminor_t minor,		/* minor device number */
   int do_write,			/* read or write? */
   u64_t position,		/* offset on device to read or write */
